Refactor `Orchestrate` so that `Create` and `Destroy` inherit from `Orchestrate`, as these two classes share mostly the same methods
